requirements_local.txt
-r requirements.txt
python-dotenv==0.14.0
pylint==2.6.0
pylint-django==2.3.0
pre-commit==2.7.1
